REFACTOR(RestApi proj): structure change
