+++ OnlineMinion.RestApi/AccountSpecsEndpoints.cs
@@ -14,7 +14,7 @@ using OnlineMinion.Contracts.Shared.Responses;
 using OnlineMinion.RestApi.Helpers;
 using OnlineMinion.RestApi.Paging;
 using static Microsoft.AspNetCore.Http.TypedResults;
-using static OnlineMinion.RestApi.CommonEndpointsValidatorUniqueByMember;
+using static OnlineMinion.RestApi.Shared.CommonValidationEndpoints;
 using static OnlineMinion.RestApi.Init.ApiCorsOptionsConfigurator;
 using static OnlineMinion.RestApi.ProblemHandling.ApiProblemsHandler;
 using static OnlineMinion.RestApi.Shared.NamedRoutes;
